seek.c をパースできるように

+++ src/seek.c
@@ -24,7 +24,7 @@ char *read_file(char *path) {
   // ファイルの長さを調べる
   if (fseek(fp, 0, SEEK_END) == -1)
     error("%s: fseek: %s", path, strerror(errno));
-  size_t size = ftell(fp);
+  int size = ftell(fp);
   if (fseek(fp, 0, SEEK_SET) == -1)
     error("%s: fseek: %s", path, strerror(errno));
